requirements.txt
aiogram==2.3
aiohttp==3.5.4
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
Babel==2.7.0
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptg==0.2.post0
cryptography==2.7
cycler==0.10.0
future==0.17.1
hachoir==3.0a5
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
matplotlib==3.1.1
mccabe==0.6.1
multidict==4.5.2
mysqlclient==1.4.4
numpy==1.17.0
pandas==0.25.1
Pillow==6.1.0
pyaes==1.6.1
pyasn1==0.4.6
pycparser==2.19
pylint==2.3.1
pyparsing==2.4.2
PySide2==5.13.0
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-telegram-bot==12.1.0
pytz==2019.2
requests==2.22.0
rope==0.14.0
rsa==4.0
selenium==3.141.0
shiboken2==5.13.0
six==1.12.0
soupsieve==1.9.3
Telethon==1.9.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
yarl==1.3.0
